feat: import Go's import path logic to assess the code repository to use
This commit adds a forked / improved version of Google's RepoRootForImportPath logic. Most of this code is originally part of the go/internal section of Google's codebase: https://github.com/golang/go/blob/16d3040a84be821d801b75bd1a3d8ab4cc89ee36/src/cmd/go/internal/vcs/vcs.go#L1281.

However, Justin found a version of this library exposed by Google without all the internal dependencies. It is now deprecated as unmaintained, but the core logic: the dynamic resolving coupled with the regexes are still relevant and enough for our use-case: https://pkg.go.dev/golang.org/x/tools/go/vcs#RepoRootForImportPath.

We had to fork the library in order to upgrade the go version used, bump the dependencies and more importantly fix a few unported changes:
1. We remove unsupported VCS, as BuildKit only support Git based VCS
2. We remove the deprecated mercurial check from bitbucket vcs options (golang/go@5b1b80b) leading to an unability to parse the root of a repo from a bitbucket private ref

However, in the context of Dagger, the important part is to be able to split the root of the repo from its subdir. I think it is ok as it will fail at the gitdns.Git level (to be verified)

3. Update the regexes and the tests for jazz, apache, git.sr.ht SCM (following current official package's logic)
4. Remove the deprecation notices, as this package does not aim to be 100% compliant with upstream but just handle the split between the root of a repo and the subpath from its URL + rely on io.ReadAll instead of ioutil

Signed-off-by: grouville <guillaume@dagger.io>