[Installation] - Improve Installation for Windows Users
- **What this feature does**:
Windows user should be able to use the dagger binary getting a similar experience to the one proposed for Linux in the docs.
- **What has been changed**: the feature implement a powershell script similar to `install.sh` logic. More specifically I've updated `install.md` and added `install.ps1`

I'm not sure if the  URL `releases.dagger.io` is currently pointing at the content of GH repo or something else so the documented steps may need a review.

Signed-off-by: alefesta <afesta@alfweb.com>