Fix data race issue by synchronizing events channel through waiting group
Previously, there was a data race condition error when `op
.#PushContainer` was called. It was because of the `Export` func
in `solver.go` that create a sub build pipeline and forward his
events to the main events channel (that catch all events that we log).
To fix it, channel are now split between build pipeline and
centralize into the old output channel (`ch`).

Signed-off-by: Tom Chauveau <tom.chauveau@epitech.eu>