Fix env & log global flags are no longer hidden.
There was a hack in `version.go` that was hiding unused flag from
that command.
The problem is that the effect has spread to all commands.

I've removed that trick to correctly display all flags because it's
important for users to know how select an environment and configure
log format.

Signed-off-by: Tom Chauveau <tom.chauveau@epitech.eu>