fix(buildkitd): Start the daemon if it is not active
Add a method that will retrieve the version, the running state and the network host existence of the buildkitd container.
If the version is outdated or there is no host network, we delete the container and install a proper one.
If the container is correctly configured but is not active, we start it, which saves some time.

Signed-off-by: Lucas Perreau <lucas.perreau@ext.adeo.com>