Default semver bump type when scheduled
When this workflow is scheduled, instead of triggered manually, we need
a default bump type, otherwise it will fail:
https://github.com/dagger/dagger/runs/4942069177?check_suite_focus=true#step:5:11

If the SEMVER_BUMP environment variable is not set via an input, which
will be the case for all scheduled runs, default to the same 'prerel
alpha..' value.

Modified the auto-trigger to later on today so that we can check the
change in the final setting.

This is a Ship & Show follow-up to
https://github.com/dagger/dagger/pull/1476

What is Ship / Show / Ask?
https://martinfowler.com/articles/ship-show-ask.html

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>