Always build the docs website
If we trigger a deploy, then we want the deploy to actually run.

I am assuming that this was in place when there was a single GitHub
Actions workflow that would trigger this all the time, even when nothing
changed in `docs/**`. Since https://github.com/dagger/dagger/pull/1498,
we no longer have this problem.

After merging https://github.com/dagger/dagger/pull/1591, I was caught
by this behaviour which to me seemed surprising. This change should make
docs deploy via Netlify behave in a more predictable way.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>