First pass at the Go on Docker Swarm doc page
There are a few more things to do before we can finish this, see the
comments in the doc. The most important item is to switch the Dagger
config to Europa. That is actually one of the next steps.

Sharing this early so that we start collaborating. Pushing to a branch
in the dagger org so that we can work on it together (personal
repositories make it more difficult).

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>