Add link to Europa docs
So that it's easy for anyone to jump to the new docs that we are
currently working on, and intend to replace the existing docs with.

While I would have preferred to link to the local dev page, it's still
stuck in the PR state, currently blocked on another PR:
https://github.com/dagger/dagger/pull/1586

Also added a link to the pre-Europa docs, so that it's easy to go back.

While at it, drop "Sidebar" from the name of sidebars, and replace
tutorial with a more descriptive name.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>