Add a DRAFTS section to the Europa docs sidebar
This is meant to act as catch-all for pages that have been written
ad-hoc, but don't currently fit anywhere.

While it was initially mentioned that this should be called
Knowledgebase, it didn't seem to accurately describe the intent.
The email & message default DRAFTS seemed better, so I just went with it.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>