Track latest version by Major.Minor when releasing via GoReleaser
We want to be able to download the latest published version for a
specific Major.Minor. In the Dagger GitHub Action, we want to be able to
restrict the versions to e.g. latest 0.2.

We took this opportunity to change the URL schema and scope all versions
to /versions, e.g. /versions/latest, /versions/0.1, /versions/0.2, etc.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>