Add CUE & continue with Dagger CUE in Europa docs sidebar
While the CUE doc page was not on my radar for the 0.2.0 release, it
became obvious that we would do our users a dis-service if we didn't add it.

The initial idea was to just include the existing page, from the 0.1.0
docs, but then we have hit a few issues and concluded that duplicating
this page with a new ID is a reasonable compromise:
https://github.com/dagger/dagger/pull/1758#issuecomment-1064417312

This was a good first step, but not good enough for the 0.2.0 docs. For
the new docs we have adopted a more direct language and slightly
different formatting:
- shorter sentences
- more lists
- and a story-like approach to encourage users to navigate from doc to doc

While our primary aim is to help users achieve specific goals (e.g. use
secrets, build container images etc.), the secondary aim is to share the
the Dagger philosophy and link different concepts into a whole which is
greater than the sum of its parts. Given this perspective, it is easy to
understand why rewriting the CUE page for 0.2.0 and linking it to the
Dagger CUE API was the right next step. That is what this change does.

As I was taking a closer look at the Dagger CUE API page, I realised
that linking to the specific 0.2.0 cue pages for dagger.io/dagger would
help users discover them and understand better CUE packages. I also
added some definitions which we have missed.

Part of this change, I have dropped the DRAFTS section since the only
page was a pre-Europa use-case. In my mind, the post-Europa docs effort
will be all about Use Cases, so it would make sense to defer this page
when we start that, instead of keeping around something
half-implemented. The cost of inventory is underrated IMO.

I have also updated the only two occurrences of 0.2.0-beta.1 to 0.2.0

This is part of https://github.com/dagger/dagger/issues/1327

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>