Add helloworld example
This is the first thing that someone new to Dagger tried to do.

I wanted to capture it in an example so that we can reference in the future.

I am sure that we can improve on this small step in a direction that I
have seen a new member take.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>