Allow any published dagger version to be installed via install.sh
This also updates Getting Started - CI/CD in your local dev - Linux & Mac doc.
A few other related doc changes were pulled in by this.

I wanted to try different versions of dagger on Linux, and this was the
easiest way of doing that:

    curl -L https://dl.dagger.io/dagger/install.sh -O
    # make this change, and then:

    DAGGER_VERSION=0.2.0 sh install.sh
    # ...
    ./bin/dagger version
    dagger 0.2.0 (e499297e) linux/amd64

    DAGGER_VERSION=0.2.1 sh install.sh
    # ...
    ./bin/dagger version
    dagger 0.2.1 (69b4845d) linux/amd64

This change enables anyone to do the following instead:

    curl -L https://dl.dagger.io/dagger/install.sh \
    | DAGGER_VERSION=0.2.0 sh

Until this gets merged, you can test how this change works in practice
via this temporary file:

    curl -L https://dl.dagger.io/dagger/install-pr-1819.sh \
    | DAGGER_VERSION=0.2.0 sh

One thing which I was not sure about is how the file from this
repository ends up on https://dl.dagger.io/dagger/install.sh. I think
this is a manual process today, but I need to confirm this assumption.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>