Start making room for the Actions API.
This commit will be rebased and the commit message changed.
I am capturing a WIP checkpoint.

My current thinking for the **entire** Core Concepts category is:
1. Start with the plan
2. Continue with the client property - it's the first thing in the plan
3. Move onto secrets - still client, but it deserves a page of its own
4. Introduce actions - the second thing in the plan
5. Reference the Actions API which shows all the core stuff
6. Re-enforce that when actions run, they create images - current
   Building container images page
7. Finish with the existing What is CUE?

I don't think that we should put anything else in here for now.

CLI Telemetry I would pull out after Use Cases & before Dagger 0.1.
It doesn't belong in any of the existing categories, so let's go for the
least amount of work for now.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>