Support for async start/stop execs.
This adds a new task type called Start which has a similar interface
as Exec, but unblocks once the exec has started rather than waiting for
it to complete. A separate type, Stop, can be used to stop it
explicitly. If it isn't stopped, then it will be killed once the dagger
client shutsdown.

This is intended for creating ephemeral services used by other tasks in
the DAG, such as a DB against which tests can be run in different tasks.

Start cannot output state and any state they create while running
will be thrown away once the DAG finishes executing. The only exception
is if they use a shared cache mount, in which case state will be applied
to the cache mount and accessible by other Execs, async or not.

There is currently an issue with Buildkit where if the dagger client
exits ungracefully without running cleanups, then the container will
leak and keep running in Buildkit. There is also no way for a subsequent
client to query what containers are running and delete any leaked ones,
so the only remedy in this case would be restarting buildkitd. This is
an issue that needs to be fixed Buildkit-side.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>