solver: fix duplicate logs on export
This is a simple fix for the issue of logs being duplicated when an
export is performed. The issue is caused by Buildkit's behavior of
"backfilling" newly connected clients on the progress so far of vertices
referenced by their build. This addresses it in the simplest way
possible by just filtering out progress log events from these separate
builds.

This works but is fragile. It works because we currently always set
Evaluate to true when running solves, which means that any progress logs
will already have been sent to the main solve channel and any that are
sent to the separate channel created for the export must be duplicates.

If we ever want to enable Evaluate being set to false, we will need a
different fix. I wrote that fix out, but it's a tangled web of
synchronization and to fully understand that it actually covers every
race-condition you need to know details about BuildKit's internal
implementation. Therefore, for now I just left a comment about all this
above the line where we set Evaluate to true and linked to the more
complicated fix in case it's ever needed.

The long-term, non-hacky fix to this is to add support for Export to the
GatewayAPI. For now though, this is our best option and is in fact what
Buildkit developers did when facing the same issue in buildx.

One last note, there technically are some logs that we may be dropping
due to this as it turns out buildkitd uses the progress log streams to
write temporary errors encountered during registry operations. I think
this is okay for now because retriable errors are a rare case and if
retries are exhausted the actual error message does get returned to the
client. We will only be missing out on some intermediary ones.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>