Use progress group to track components of vertices. (#2306)
This change replaces the use of "@<component name>@-<name>" for tracking
which cue fields progress status updates from Buildkit should be logged
with. Now, the ProgressGroup metadata field is used to track the
component part (i.e. cue field) and the vertex name can just be the name
again.

This fixes the issue where Dockerfile builds showed up under the
"system" group. It also is more performant as statuses don't all need to
be parsed anymore.

The only exception that remains are the progress logs written for
ResolveImageConfig calls. This Buildkit API is missing support for
ProgressGroup right now, so it still uses the '@%s@' mechanism as
before. This can be fixed once Buildkit has been updated upstream to
support the field.

This exception includes our own call to ResolveImageConfig but also one
done by the Dockerfile frontend. For this reason, some of the previous
changes that have special handling for logging the Dockerfile task are
left in place for now.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>