cache: Error out if GHA used w/out token or url.
Before this change, we were using code from the buildx repo that
silently removed GHA from cache options if the GHA token and/or url were
not set. This made it very non-obvious that a mistake was being made,
with the only evidence being that cache import/export just never ran.

This commit updates to use the buildctl code for parsing cache options
(which is better since it is tied to the version of buildkit we are
using, unlike buildx). We need to layer the GHA token+url checks on top
ourselves because buildctl lacks those checks, but that's fairly
straightforward.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>