Revert specific case to handle dockerfile actions
After the changes introduced by sipsma in
https://github.com/dagger/dagger/commit/86e798a36bc93d19ac33631d0cb466ab771fa164
Dockerfile logs are now displaying correctly and don't require the
system hack. Additionally, I'm removing the CanTransition method and
going back to the previous behavior since it's not required anymore and
it also introduces a bug when displaying logs for actions that contain
hidden sub-actions.

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>