docs: Migrate from docsearch-scraper to https://docsearch.algolia.com
Instead of continuing with the now deprecated docsearch-scraper, we
should instead migrate to DocSearch, as mentioned in this blog post:
https://docusaurus.io/blog/2021/11/21/algolia-docsearch-migration

These are the next steps that I can see:
- [x] Apply for DocSearch → https://support.algolia.com/hc/en-us/requests/491210
- [ ] Open a new PR which shows our DocSearch migration intent in our public repo
- [ ] Update support request with PR details
- [ ] ...

My expectation is that we will figure out the remaining steps as we get
remaining perspectives involved. I'm thinking Jean-François & Clément.

For now, this is enough to kick it off ⚽️

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>