client: use lock file to dedupe buildkitd creation.
Before this, there was a race condition where parallel dagger clients
could race while checking+starting buildkitd, which sometimes lead to
one removing the buildkitd container created by the other.

Now a lock file is used to ensure the process of checking+starting
buildkitd is transactional.

This doesn't fix the case where dockerd is remote and clients are
running on parallel machines, but that's more obscure and much tougher
to fix as dockerd doesn't appear to have any mechanisms for idempotently
creating a container.

The lock file is under ~/.config/dagger for now. Technically, the
correct place to put this according to the FHS would be /var/lock, but
in my experience /var/lock doesn't always exist and I want to avoid
having to create a directory under /var if possible for now.

Also, there is a long timeout on waiting to acquire the lock. This is
set in order to avoid code that could block for infinity, but is quite
long to account for any users that are trying to pull the buildkitd
image on a particularly slow internet connection.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>