Improve start/stop test.
This addresses an issue where the start step execParamsTest was using a
constant cache mount id which was also used by a totally different test.
This made it possible for these tests to interfere with each other.

This also fixes an issue where because we were setting "always" to true
in the sync exec, if the sync exec started before the async one, it would
end up with a not-actually-shared-mount (due to some bizarre but
expected behavior in BuildKit).

The commit fixes all of the above by using a random ID for the cache
mount, which ensures it doesn't overlap with other test cases and
removes the need for setting always to true.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>