Downgrade opentelemetry to match Buildkit's version.
Before this, when trying to use opentelemetry libs it was possible to
get an error "cannot merge resource due to conflicting Schema URL".

The error indicates that different versions of opentelemetry are being
used at the same time. This seems to have ultimately resulted from the
fact that we setup tracers and also provide those to some Buildkit libs,
but buildkit is still on v1.4.1 while we recently upgraded to v1.7.0.

This, for reasons that are not yet clear to me, only happened in the
integration tests where the dagger client is running inside dagger. It
may have something to do with a difference in how we build dagger inside
dagger compared to when it is built directly in GHA, but not sure.

Either way, this downgrade unblocks us for the moment until we have a
better way of solving this.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>