ci: add tracing support to integration tests.
This sets up Jaeger to run inside GHA during integration tests,
instruments dagger and buildkit to write to it and exports the resulting
data as an artifact of the run.

This enables us to get insight into what actually happened during our CI
runs, which is helpful for debugging and getting insight into the
performance of caching.

Jaeger seems to be mostly designed for "live" use, but it has an
experimental mode for persisting its data to disk, which I use here to
enable exporting it.

This is not a completely ideal solution. It requires setting up Buildkit
and Jaeger in docker separate from our actual test run. The much better
solution will be to use async \#Start and \#Stop tasks to run at least
Jaeger (and probably Buildkit itself too). This will allows use to
dogfood our own features and also open the interesting possibility of
making it easier for other users to setup Jaeger or other
opentelemetry-related daemons too (they can just re-use their Buildkit
infra instead of setting up something else).

\#Start will need a few more features first though, the main one being
support for sending SIGTERM before SIGKILL (right now it is just
instantly SIGKILL'd). Without that, the Jaeger data is corrupted. Once
that support is merged, we will be able to switch over to \#Start.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>