ci: add plan for viewing Jaeger CI traces locally
This allows you to run a single dagger do command that pulls down jaeger
data from the artifacts of a GHA run, start a local Jaeger daemon with
it and then view it in your web browser.

The dagger do command for this is printed at the end of GHA runs with
all the --with parameters filled out, for convenience.

It uses GH credentials, which it assumes you have previously setup on
your local machine via "gh auth login" (gh is the github cli).

This would be a great use-case for client-side proxying, but that
doesn't exist yet, so the Jaeger daemon will just end up listening on
the localhost of the machine running Buildkit. This is better than
nothing for now and use of SSH port forwarding can help cases where its
otherwise convenient.

It is also a bit weird in that this is a plan that is expected to just
keep running indefinitely until you cancel it (e.g. with ctrl-C).
However, I have not experienced any issues with the previously, other
than a weirdly flashing tty output while its running (which is a
separate issue).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>