Add gcp serverless cloud function package
This package is named under gcp in order for it to later be completed
with more gcp features.
There is for now still a bug that appeared recently and wasn't present
like a month ago. It needs to be fixed, I just can't figure out how for
now and I'll appreciate any help.

Signed-off-by: Benjamin Reigner <benjamin.reigner@epitech.eu>