Support signaling core.Start execs.
This adds a new core.SendSignal task that enables arbitrary signals to
be sent to an async Start exec. It also adds related support to
core.Stop for specifying a timeout to wait for the exec to exit on its
own before sending SIGKILL.

The main goal here is to support ending Start execs gracefully instead
of being forced to always just SIGKILL them as part of core.Stop. It
should be useful in general for any other use cases that call for
signals to be sent to these execs (i.e. SIGWINCH for future tty
support).

Among other things, this is needed to support running Jaeger via
core.Start (without a graceful stop, its data gets corrupted).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>