gitpod: mark /src directory as safe to use
More recent versions of git include extra safety checks on the owner of
directories before doing any operations on them. This meant that various
calls to git would fail due to being copied from the host, where the
source code is most often not owned by root (the user in the container).

This fixes the issue by just telling git it's okay to use the /src
directory anyways.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>