ci: use prod dagger binary in universe tests
For reasons that are not clear to me at this time, when using the
dagger-debug binary to run universe tests, the new yarn test is
incredibly slow and usually times out after 30 minutes. But when using
the default dagger binary, it's relatively fast.

This needs to be understood better, but for now, running the universe
tests with the default binary will unblock our CI (which currently fails
on this yarn test consistently).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>