Send log events async using TelemetryLite
The most important thing which needs improving is handling re-auth
correctly. For now we just panic, but I suspect that Context is part of
it, as is the passing of TelemetryLite all the way into the Cloud
logger. There is a much better way of structuring this, but this will
have to do for now.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>