Use a worker pool to send telemetry data
This speeds up telemetry upload considerably. This is an initial
approach before moving into more advanced solutions like using http2
streams or forking the process to handle this in the background. So far
the performance has been acceptable, so we can say it's good enough for
a simple solution for the moment.

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>