universe(go): revert of part of #2632
This commit revert a change of API in `universe/go`.
A recent PR added 2 fields: `name` and `repository` to customize the image to
pull when calling `go.#Image`.
The problem is that change does not fit with the purpose of `go.#Image`, this
one should be used as default image to cover simple workflow with the simplest
possible API. That's why it's usually only possible to configure some packages
and a version of the official image.

If the user want to provide a custom image, he can override the field `input`
(or image sometime) of other definition.
To do so, he can use `docker.#Pull` or create his own definition to cover his
proper use case.

Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>