chore: Bump Go to 1.18
GoReleaser v1.10 update broke our releases:
https://github.com/dagger/dagger/runs/7205381884?check_suite_focus=true#step:8:116

It tried building a windows_arm64 version which is not possible with Go
1.16. This has more details:
https://github.com/goreleaser/goreleaser/issues/3224

Instead of pinning to GoReleaser v1.9.2 (last version that worked
https://github.com/dagger/dagger/runs/7097754532?check_suite_focus=true#step:8:15)
we should do the right thing and bump the Go version that we are using
to develop Dagger to 1.18. By the way, some of us are already using 1.18
locally, but not CI, so this change fixes it. It should also fix the
release.

Resolves https://github.com/dagger/dagger/issues/1859

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>