build(deps): Bump jaeger dep from 1.4.1 to 1.8.0
This is similar to https://github.com/dagger/dagger/pull/2766, but with
a GHA cache fix. Integration tests pass consistently locally, on Linux,
but fail consistently on GitHub Actions. My only explanation is the
GitHub Actions cache. This is the follow-up to this comment:
https://github.com/dagger/dagger/pull/2766#issuecomment-1183851791

FWIW, I expect integration tests to fail for this commit, same as
https://github.com/dagger/dagger/runs/7331502142?check_suite_focus=true

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>