ci: Use go.mod replaced CUE version for integration tests (#2801)
Rather than using first version of CUE - head -1 - from go.mod, we want
to use the last version - tail -n 1. While I am not sure why we have two
versions of CUE in the go.mod, based on various internal discussions, I
do know that there are various edge case issues with CUE > 0.4.1. With
the current progress that upstream contributions are making, we continue
being stuck with 0.4.1 for now. Related to:
- https://github.com/dagger/dagger/issues/2182
- https://github.com/dagger/dagger/issues/2318

While at it, replace grep | cut | sed with a simpler awk expression. It
splits on either 'v' or '-' and uses the second element, i.e. '0.4.3' &
'0.4.1'. We end up using the last CUE version from go.mod, 0.4.1 in this case.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>